Recreate node on a reconnection.

+++ src/main/java/com/librato/disco/DiscoService.java
@@ -3,7 +3,10 @@ package com.librato.disco;
 import com.google.common.base.Preconditions;
 import org.apache.curator.framework.CuratorFramework;
 import org.apache.curator.framework.imps.CuratorFrameworkState;
+import org.apache.curator.framework.state.ConnectionState;
+import org.apache.curator.framework.state.ConnectionStateListener;
 import org.apache.zookeeper.CreateMode;
+import org.apache.zookeeper.KeeperException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -18,6 +21,7 @@ public class DiscoService {
     final String nodeName;
     final int port;
     final String node;
+    final ConnectionStateListener listener;
 
     public DiscoService(CuratorFramework framework, String serviceName, String nodeName, int port) {
         this.framework = framework;
@@ -28,6 +32,24 @@ public class DiscoService {
         // Register ephemeral node as representation of this service's nodename and port
         // such as /services/myservice/nodes/192.168.1.1:8000
         this.node = baseNode + "/" + nodeName + ":" + port;
+
+        this.listener = new ConnectionStateListener() {
+            @Override
+            public void stateChanged(CuratorFramework curatorFramework, ConnectionState connectionState) {
+                if (connectionState == ConnectionState.RECONNECTED) {
+                    log.info("Re-registering with ZK as node {}", node);
+
+                    try {
+                        deleteNode();
+                        createNode();
+                    } catch (Exception e) {
+                        log.error("Exception recreating path", e);
+                        throw new RuntimeException(e);
+                    }
+                }
+            }
+        };
+
     }
 
     public void start() throws Exception {
@@ -41,14 +63,27 @@ public class DiscoService {
                     .forPath(baseNode);
         }
 
-        // TODO: Add payload capability
         log.info("Registering with ZK as node {}", node);
+        createNode();
+
+        framework.getConnectionStateListenable().addListener(listener);
+    }
+
+    public void stop() throws Exception {
+        framework.getConnectionStateListenable().removeListener(listener);
+        deleteNode();
+    }
+
+    private void createNode() throws Exception {
+        // TODO: Add payload capability
         framework.create()
                 .withMode(CreateMode.EPHEMERAL)
                 .forPath(node);
     }
 
-    public void stop() throws Exception {
-        framework.delete().forPath(node);
+    private void deleteNode() throws Exception {
+        if (framework.checkExists().forPath(node) != null) {
+            framework.delete().forPath(node);
+        }
     }
 }
